Clear code and seperate classes
